FB-13-fix 기초 엔티티 수정

+++ src/main/java/mono/focusider/domain/quiz/repository/KeywordRepository.java
@@ -3,5 +3,5 @@ package mono.focusider.domain.quiz.repository;
 import mono.focusider.domain.quiz.domain.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeywordRepository extends JpaRepository<Keyword, Integer> {
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
 }
